origttl should be a long, not an int, in the constructors.
git-svn-id: http://svn.code.sf.net/p/dnsjava/code/trunk@1198 c76caeb1-94fd-44dd-870f-0c9d92034fc1

+++ org/xbill/DNS/RRSIGRecord.java
@@ -48,7 +48,7 @@ getMember() {
  * @param signature Binary data representing the signature
  */
 public
-RRSIGRecord(Name name, int dclass, long ttl, int covered, int alg, int origttl,
+RRSIGRecord(Name name, int dclass, long ttl, int covered, int alg, long origttl,
 	    Date expire, Date timeSigned, int footprint, Name signer,
 	    byte [] signature)
 {
